controller, vmi: Extract NIC hotplug logic
Following PR #6852, split the NIC hotplug logic to a separate file.
